# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HP PrintOS Site Flow API",
    "description": "API for managing print orders, including validation, submission, and status tracking.",
    "version": "1.0.0",
    "contact": {
      "name": "HP Developer Portal",
      "url": "https://developers.hp.com/"
    }
  },
  "servers": [
    {
      "url": "https://printos.api.hp.com/siteflow",
      "description": "Production server (for existing PrintOS customers)"
    },
    {
      "url": "https://stage.printos.api.hp.com/siteflow",
      "description": "Staging server (for 3rd parties)"
    }
  ],
  "paths": {
...
    "/api/order": {
      "post": {
        "summary": "Submit an Order",
        "description": "Submits a new print order.\n\nThe order payload is uploaded to OSS as a JSON file and a pre-signed download URL is included in the response.\n\n**Duplicate detection:** If an order with the same `sourceOrderId` already exists for this store, the API returns HTTP **400** with a SiteFlow-compatible error response.",
        "operationId": "submitOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSubmissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — duplicate sourceOrderId detected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCreationErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "ofError": true,
                    "statusCode": 400,
                    "code": 208,
                    "message": "Validation Failed",
                    "validations": [
                      {
                        "path": "orderData.sourceOrderId",
                        "message": "Source Order ID already exists"
                      }
                    ],
                    "mongoErr": true
                  }
                }
              }
            }
          },
          "422": {
            "description": "Request body validation failed (missing required fields, wrong types, etc.).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCreationErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "ofError": true,
                    "statusCode": 422,
                    "message": "Validation Failed",
                    "validations": [
                      {
                        "path": "orderData.sourceOrderId",
                        "message": "Field required"
                      }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error during order submission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCreationErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "ofError": true,
                    "statusCode": 500,
                    "message": "Order submission failed: internal error"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get All Orders",
        "description": "Retrieves a paginated list of orders.",
        "operationId": "getAllOrders",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "The page number to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "description": "Number of orders per page.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderSubmissionRequest": {
        "type": "object",
        "properties": {
          "destination": {
            "$ref": "#/components/schemas/Destination"
          },
          "orderData": {
            "$ref": "#/components/schemas/OrderData"
          },
          "source": {
            "type": "object",
            "description": "Source information."
          },
          "files": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Associated files."
          }
        }
      },
      "OrderData": {
        "type": "object",
        "properties": {
          "sourceOrderId": {
            "type": "string"
          },
          "postbackAddress": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "shipments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Shipment"
            }
          },
          "stockItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockItem"
            },
            "description": "Stock items required for production."
          },
          "printType": {
            "type": "string",
            "description": "Print type (e.g., 'digital')."
          },
          "email": {
            "type": "string",
            "description": "Customer email."
          },
          "amount": {
            "type": "number",
            "description": "Order amount."
          },
          "customerName": {
            "type": "string",
            "description": "Customer name."
          },
          "extraData": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Extra order data."
          },
          "status": {
            "type": "string",
            "description": "Order status (read-only, returned in responses only). Mapping: received/pending→received, validated/processing→dataready, printready→printready, printed→printed, shipped→shipped, errored/failed→error, cancelled→cancelled.",
            "enum": ["received", "dataready", "printready", "printed", "shipped", "error", "cancelled"],
            "readOnly": true
          }
        }
      },
      "StockItem": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stock item code."
          },
          "quantity": {
            "type": "integer",
            "minimum": 0,
            "description": "Stock quantity."
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "sourceItemId": {
            "oneOf": [
              {"type": "string"},
              {"type": "integer"}
            ],
            "description": "External item ID."
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "printQuantity": {
            "type": "integer",
            "minimum": 1
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          },
          "unitCost": {
            "type": "number",
            "minimum": 0
          },
          "unitWeight": {
            "type": "number",
            "minimum": 0
          },
          "description": {
            "type": "string",
            "description": "Item description."
          },
          "productDescription": {
            "type": "string",
            "description": "Product description (alias for description)."
          },
          "pages": {
            "type": "integer",
            "description": "Number of pages."
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "shipmentIndex": {
            "type": "integer",
            "description": "Index linking this item to a shipment."
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Component"
            }
          },
          "extraData": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Extra item data."
          }
        }
      },
      "Component": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "fetch": {
            "type": "boolean",
            "default": true
          },
          "path": {
            "type": "string"
          },
          "localFile": {
            "type": "boolean",
            "description": "Whether the file is local."
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          },
          "paperId": {
            "type": "string"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          },
          "colour": {
            "$ref": "#/components/schemas/Color"
          },
          "finish": {
            "$ref": "#/components/schemas/Finish"
          },
          "extraData": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Extra component data."
          }
        }
      },
      "Color": {
        "type": "object",
        "properties": {
          "side1": {
            "type": "string",
            "description": "Front side color."
          },
          "side2": {
            "type": "string",
            "description": "Back side color."
          }
        }
      },
      "Finish": {
        "type": "object",
        "properties": {
          "side1": {
            "type": "string",
            "description": "Front side finish."
          },
          "side2": {
            "type": "string",
            "description": "Back side finish."
          }
        }
      },
      "Shipment": {
        "type": "object",
        "properties": {
          "shipTo": {
            "$ref": "#/components/schemas/Address"
          },
          "carrier": {
            "$ref": "#/components/schemas/Carrier"
          },
          "shipmentIndex": {
            "type": "integer",
            "description": "Index linking items to this shipment."
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Shipment attachments."
          },
          "pspBranding": {
            "type": "boolean",
            "description": "PSP branding flag."
          },
          "returnAddress": {
            "$ref": "#/components/schemas/Address"
          }
        }
      },
      "Carrier": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Carrier code (e.g., fedex)."
          },
          "service": {
            "type": "string",
            "description": "Service level (e.g., ground)."
          },
          "serviceId": {
            "type": "string",
            "description": "Service identifier."
          },
          "alias": {
            "type": "string",
            "description": "Carrier alias (e.g., 'tracked')."
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Recipient name."
          },
          "companyName": {
            "type": "string",
            "description": "Company name."
          },
          "address1": {
            "type": "string",
            "description": "Street address line 1."
          },
          "address2": {
            "type": "string",
            "description": "Street address line 2."
          },
          "address3": {
            "type": "string",
            "description": "Street address line 3."
          },
          "town": {
            "type": "string",
            "description": "City/Town."
          },
          "state": {
            "type": "string",
            "description": "State/Province."
          },
          "postcode": {
            "type": "string",
            "description": "Postal code."
          },
          "isoCountry": {
            "type": "string",
            "description": "ISO country code (e.g., US)."
          },
          "country": {
            "type": "string",
            "description": "Country name (full text, e.g., 'Japan')."
          },
          "email": {
            "type": "string",
            "description": "Contact email."
          },
          "phone": {
            "type": "string",
            "description": "Contact phone."
          }
        }
      },
...
      "OrderSubmissionResponse": {
        "type": "object",
        "description": "SiteFlow-compatible order creation response.",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Internal order ID."
          },
          "url": {
            "type": "string",
            "description": "Pre-signed OSS URL to the uploaded order payload JSON."
          },
          "timestamp": {
            "type": "string",
            "description": "ISO-8601 timestamp of order creation.",
            "example": "2026-08-11T16:04:46.131Z"
          },
          "sourceAccountId": {
            "type": "string",
            "description": "Base64-encoded store_id of the client."
          }
        }
      },
...
      "OrdersListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "count": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderSummary"
            }
          }
        }
      },
      "OrderSummary": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "destination": {
            "type": "object"
          },
          "source": {
            "type": "object"
          },
          "orderData": {
            "type": "object"
          }
        }
      },
...
      "Destination": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Account name for order destination."
          }
        }
      },
...
      "OrderCreationErrorResponse": {
        "type": "object",
        "description": "SiteFlow-compatible error response for order creation failures.",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Always false for errors.",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "ofError": {
                "type": "boolean",
                "description": "Whether this is an OneFlow error.",
                "example": true
              },
              "statusCode": {
                "type": "integer",
                "description": "HTTP status code."
              },
              "code": {
                "type": "integer",
                "description": "Internal error code."
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message.",
                "example": "Validation Failed"
              },
              "validations": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string",
                      "description": "Dotted path to the invalid field."
                    },
                    "message": {
                      "type": "string",
                      "description": "Validation error message."
                    }
                  }
                }
              },
              "mongoErr": {
                "type": "boolean",
                "description": "Whether this is a MongoDB error."
              }
            }
          }
        }
      },
...
    }
  }
}